// src/app/models/address.ts
export interface Address {
  id?: string;
  nameOfRecipient?: string;
  exact?: string;
  provinceId?: string;
  provinceName?: string;
  districtId?: string;
  districtName?: string;
  wardCode?: string;
  wardName?: string;
  combination?: string;
  phoneNumber?: string;
  userId?: string;
  isDefault?: boolean;
  active?: boolean;
}